Use double instead of float

+++ src/TiledMatrix.c
@@ -6,7 +6,7 @@
 #define GET_M(self) \
   (self->m_blk * self->side_blk)
 
-float * TiledMatrix_get_block(TiledMatrix* self,
+double * TiledMatrix_get_block(TiledMatrix* self,
                               int i_blk, int j_blk)
 {
   return self->data + i_blk * self->m_blk * self->side_blk * self->side_blk +
@@ -15,14 +15,14 @@ float * TiledMatrix_get_block(TiledMatrix* self,
 
 void TiledMatrix_set_at(TiledMatrix* self,
                         int i_blk, int j_blk,
-                        int i, int j, float val)
+                        int i, int j, double val)
 {
   self->data[i_blk * self->m_blk * self->side_blk * self->side_blk +
              j_blk * self->side_blk * self->side_blk +
              i * self->side_blk + j] = val;
 }
 
-float TiledMatrix_get_at(TiledMatrix* self,
+double TiledMatrix_get_at(TiledMatrix* self,
                          int i_blk, int j_blk,
                          int i, int j)
 {
@@ -31,7 +31,7 @@ float TiledMatrix_get_at(TiledMatrix* self,
                     i * self->side_blk + j];
 }
 
-float TiledMatrix_get_val_non_tiled(TiledMatrix* self,
+double TiledMatrix_get_val_non_tiled(TiledMatrix* self,
                                     int i, int j)
 {
   int blk_i, blk_j, pos_i, pos_j;
@@ -56,7 +56,7 @@ void TiledMatrix_print(TiledMatrix* self, FILE *output)
   for (i = 0; i < n; i++) {
     for (j = 0; j < m; j++) {
       const char* format;
-      float val;
+      double val;
 
       format = (j == m - 1) ? "%f\n" : "%f ";
       val = TiledMatrix_get_val_non_tiled(self, i, j);
